@@ -8,11 +8,9 @@
   },
   "repository": "https://github.com/Tyriar/node-pty",
   "author": "Tyriar",
-  "devDependencies": {
-    "electron": "1.8.4"
-  },
   "dependencies": {
-    "node-pty": "^0.7.4",
-    "xterm": "^3.6.0"
+    "electron": "4.0.1",
+    "node-pty": "^0.9.0",
+    "xterm": "4.2.0"
   }
 }
