@@ -2,9 +2,9 @@
   "name": "node-pty",
   "description": "Fork pseudoterminals in Node.JS",
   "author": {
-    "name": "Daniel Imms"
+    "name": "Microsoft Corporation"
   },
-  "version": "0.7.6",
+  "version": "0.9.0",
   "license": "MIT",
   "main": "./lib/index.js",
   "types": "./typings/node-pty.d.ts",
@@ -34,23 +34,25 @@
   ],
   "scripts": {
     "tsc": "tsc",
+    "watch": "tsc -w",
     "lint": "tslint 'src/**/*.ts'",
     "install": "node scripts/install.js",
     "postinstall": "node scripts/post-install.js",
     "test": "cross-env NODE_ENV=test mocha -R spec --exit lib/*.test.js",
-    "prepublish": "npm run tsc"
+    "prepare": "npm run tsc",
+    "prepublishOnly": "npm run tsc"
   },
   "dependencies": {
-    "nan": "2.10.0"
+    "nan": "^2.14.0"
   },
   "devDependencies": {
-    "@types/mocha": "^5.0.0",
-    "@types/node": "^6.0.104",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "8",
     "cross-env": "^5.1.4",
-    "mocha": "^5.0.5",
-    "pollUntil": "^1.0.3",
-    "tslint": "^5.9.1",
-    "tslint-consistent-codestyle": "^1.13.0",
-    "typescript": "^2.8.1"
+    "mocha": "^7.1.1",
+    "ps-list": "^6.0.0",
+    "tslint": "5.12.1",
+    "tslint-consistent-codestyle": "^1.15.0",
+    "typescript": "3.4"
   }
 }
