@@ -1,15 +1,25 @@
 /**
  * Copyright (c) 2012-2015, Christopher Jeffrey (MIT License)
  * Copyright (c) 2016, Daniel Imms (MIT License).
+ * Copyright (c) 2018, Microsoft Corporation (MIT License).
  */
 
-import * as path from 'path';
 import { Socket } from 'net';
 import { EventEmitter } from 'events';
 import { ITerminal, IPtyForkOptions } from './interfaces';
+import { EventEmitter2, IEvent } from './eventEmitter2';
+import { IExitEvent } from './types';
 
 export const DEFAULT_COLS: number = 80;
 export const DEFAULT_ROWS: number = 24;
+
+/**
+ * Default messages to indicate PAUSE/RESUME for automatic flow control.
+ * To avoid conflicts with rebound XON/XOFF control codes (such as on-my-zsh),
+ * the sequences can be customized in `IPtyForkOptions`.
+ */
+const FLOW_CONTROL_PAUSE =  '\x13';   // defaults to XOFF
+const FLOW_CONTROL_RESUME = '\x11';   // defaults to XON
 
 export abstract class Terminal implements ITerminal {
   protected _socket: Socket;
@@ -26,8 +36,18 @@ export abstract class Terminal implements ITerminal {
   protected _writable: boolean;
 
   protected _internalee: EventEmitter;
+  private _flowControlPause: string;
+  private _flowControlResume: string;
+  public handleFlowControl: boolean;
+
+  private _onData = new EventEmitter2<string>();
+  public get onData(): IEvent<string> { return this._onData.event; }
+  private _onExit = new EventEmitter2<IExitEvent>();
+  public get onExit(): IEvent<IExitEvent> { return this._onExit.event; }
 
   public get pid(): number { return this._pid; }
+  public get cols(): number { return this._cols; }
+  public get rows(): number { return this._rows; }
 
   constructor(opt?: IPtyForkOptions) {
     // for 'close'
@@ -39,18 +59,59 @@ export abstract class Terminal implements ITerminal {
 
     // Do basic type checks here in case node-pty is being used within JavaScript. If the wrong
     // types go through to the C++ side it can lead to hard to diagnose exceptions.
-    this._checkType('name', opt.name ? opt.name : null, 'string');
-    this._checkType('cols', opt.cols ? opt.cols : null, 'number');
-    this._checkType('rows', opt.rows ? opt.rows : null, 'number');
-    this._checkType('cwd', opt.cwd ? opt.cwd : null, 'string');
-    this._checkType('env', opt.env ? opt.env : null, 'object');
-    this._checkType('uid', opt.uid ? opt.uid : null, 'number');
-    this._checkType('gid', opt.gid ? opt.gid : null, 'number');
-    this._checkType('encoding', opt.encoding ? opt.encoding : null, 'string');
+    this._checkType('name', opt.name ? opt.name : undefined, 'string');
+    this._checkType('cols', opt.cols ? opt.cols : undefined, 'number');
+    this._checkType('rows', opt.rows ? opt.rows : undefined, 'number');
+    this._checkType('cwd', opt.cwd ? opt.cwd : undefined, 'string');
+    this._checkType('env', opt.env ? opt.env : undefined, 'object');
+    this._checkType('uid', opt.uid ? opt.uid : undefined, 'number');
+    this._checkType('gid', opt.gid ? opt.gid : undefined, 'number');
+    this._checkType('encoding', opt.encoding ? opt.encoding : undefined, 'string');
+
+    // setup flow control handling
+    this.handleFlowControl = !!(opt.handleFlowControl);
+    this._flowControlPause = opt.flowControlPause || FLOW_CONTROL_PAUSE;
+    this._flowControlResume = opt.flowControlResume || FLOW_CONTROL_RESUME;
   }
 
-  private _checkType(name: string, value: any, type: string): void {
-    if (value && typeof value !== type) {
+  protected abstract _write(data: string): void;
+
+  public write(data: string): void {
+    if (this.handleFlowControl) {
+      // PAUSE/RESUME messages are not forwarded to the pty
+      if (data === this._flowControlPause) {
+        this.pause();
+        return;
+      }
+      if (data === this._flowControlResume) {
+        this.resume();
+        return;
+      }
+    }
+    // everything else goes to the real pty
+    this._write(data);
+  }
+
+  protected _forwardEvents(): void {
+    this.on('data', e => this._onData.fire(e));
+    this.on('exit', (exitCode, signal) => this._onExit.fire({ exitCode, signal }));
+  }
+
+  protected _checkType<T>(name: string, value: T | undefined, type: string, allowArray: boolean = false): void {
+    if (value === undefined) {
+      return;
+    }
+    if (allowArray) {
+      if (Array.isArray(value)) {
+        value.forEach((v, i) => {
+          if (typeof v !== type) {
+            throw new Error(`${name}[${i}] must be a ${type} (not a ${typeof v[i]})`);
+          }
+        });
+        return;
+      }
+    }
+    if (typeof value !== type) {
       throw new Error(`${name} must be a ${type} (not a ${typeof value})`);
     }
   }
@@ -76,7 +137,7 @@ export abstract class Terminal implements ITerminal {
   }
 
   /** See net.Socket.setEncoding */
-  public setEncoding(encoding: string): void {
+  public setEncoding(encoding: string | null): void {
     if ((<any>this._socket)._decoder) {
       delete (<any>this._socket)._decoder;
     }
@@ -117,7 +178,6 @@ export abstract class Terminal implements ITerminal {
     this._socket.once(eventName, listener);
   }
 
-  public abstract write(data: string): void;
   public abstract resize(cols: number, rows: number): void;
   public abstract destroy(): void;
   public abstract kill(signal?: string): void;
@@ -125,19 +185,6 @@ export abstract class Terminal implements ITerminal {
   public abstract get process(): string;
   public abstract get master(): Socket;
   public abstract get slave(): Socket;
-
-  // TODO: Should this be in the API?
-  public redraw(): void {
-    let cols = this._cols;
-    let rows = this._rows;
-
-    // We could just send SIGWINCH, but most programs will  ignore it if the
-    // size hasn't actually changed.
-
-    this.resize(cols + 1, rows + 1);
-
-    setTimeout(() => this.resize(cols, rows), 30);
-  }
 
   protected _close(): void {
     this._socket.writable = false;
