@@ -1,14 +1,13 @@
 /**
  * Copyright (c) 2012-2015, Christopher Jeffrey, Peter Sunde (MIT License)
  * Copyright (c) 2016, Daniel Imms (MIT License).
+ * Copyright (c) 2018, Microsoft Corporation (MIT License).
  */
 
-import * as path from 'path';
 import { Socket } from 'net';
-import { inherits } from 'util';
 import { Terminal, DEFAULT_COLS, DEFAULT_ROWS } from './terminal';
 import { WindowsPtyAgent } from './windowsPtyAgent';
-import { IPtyForkOptions, IPtyOpenOptions } from './interfaces';
+import { IPtyOpenOptions, IWindowsPtyForkOptions } from './interfaces';
 import { ArgvOrCommandLine } from './types';
 import { assign } from './utils';
 
@@ -20,8 +19,10 @@ export class WindowsTerminal extends Terminal {
   private _deferreds: any[];
   private _agent: WindowsPtyAgent;
 
-  constructor(file?: string, args?: ArgvOrCommandLine, opt?: IPtyForkOptions) {
+  constructor(file?: string, args?: ArgvOrCommandLine, opt?: IWindowsPtyForkOptions) {
     super(opt);
+
+    this._checkType('args', args, 'string', true);
 
     // Initialize arguments
     args = args || [];
@@ -34,8 +35,8 @@ export class WindowsTerminal extends Terminal {
     }
 
     const env = assign({}, opt.env);
-    const cols = opt.cols || DEFAULT_COLS;
-    const rows = opt.rows || DEFAULT_ROWS;
+    this._cols = opt.cols || DEFAULT_COLS;
+    this._rows = opt.rows || DEFAULT_ROWS;
     const cwd = opt.cwd || process.cwd();
     const name = opt.name || env.TERM || DEFAULT_NAME;
     const parsedEnv = this._parseEnv(env);
@@ -47,7 +48,7 @@ export class WindowsTerminal extends Terminal {
     this._deferreds = [];
 
     // Create new termal.
-    this._agent = new WindowsPtyAgent(file, args, parsedEnv, cwd, cols, rows, false);
+    this._agent = new WindowsPtyAgent(file, args, parsedEnv, cwd, this._cols, this._rows, false, opt.useConpty, opt.conptyInheritCursor);
     this._socket = this._agent.outSocket;
 
     // Not available until `ready` event emitted.
@@ -61,7 +62,7 @@ export class WindowsTerminal extends Terminal {
 
       // These events needs to be forwarded.
       ['connect', 'data', 'end', 'timeout', 'drain'].forEach(event => {
-        this._socket.on(event, data => {
+        this._socket.on(event, () => {
 
           // Wait until the first data event is fired then we can run deferreds.
           if (!this._isReady && event === 'data') {
@@ -107,7 +108,7 @@ export class WindowsTerminal extends Terminal {
 
       // Cleanup after the socket is closed.
       this._socket.on('close', () => {
-        this.emit('exit', this._agent.getExitCode());
+        this.emit('exit', this._agent.exitCode);
         this._close();
       });
 
@@ -118,6 +119,16 @@ export class WindowsTerminal extends Terminal {
 
     this._readable = true;
     this._writable = true;
+
+    this._forwardEvents();
+  }
+
+  protected _write(data: string): void {
+    this._defer(this._doWrite, data);
+  }
+
+  private _doWrite(data: string): void {
+    this._agent.inSocket.write(data);
   }
 
   /**
@@ -129,25 +140,17 @@ export class WindowsTerminal extends Terminal {
   }
 
   /**
-   * Events
-   */
-
-  public write(data: string): void {
-    this._defer(() => {
-      this._agent.inSocket.write(data);
-    });
-  }
-
-  /**
    * TTY
    */
 
   public resize(cols: number, rows: number): void {
-    if (cols <= 0 || rows <= 0) {
+    if (cols <= 0 || rows <= 0 || isNaN(cols) || isNaN(rows) || cols === Infinity || rows === Infinity) {
       throw new Error('resizing must be done using positive cols and rows');
     }
     this._defer(() => {
       this._agent.resize(cols, rows);
+      this._cols = cols;
+      this._rows = rows;
     });
   }
 
@@ -167,22 +170,16 @@ export class WindowsTerminal extends Terminal {
     });
   }
 
-  private _defer(deferredFn: Function): void {
-
-    // Ensure that this method is only used within Terminal class.
-    if (!(this instanceof WindowsTerminal)) {
-      throw new Error('Must be instanceof WindowsTerminal');
-    }
-
+  private _defer<A extends any>(deferredFn: (arg?: A) => void, arg?: A): void {
     // If the terminal is ready, execute.
     if (this._isReady) {
-      deferredFn.apply(this, null);
+      deferredFn.call(this, arg);
       return;
     }
 
     // Queue until terminal is ready.
     this._deferreds.push({
-      run: () => deferredFn.apply(this, null)
+      run: () => deferredFn.call(this, arg)
     });
   }
 
